Test and fix migration up to Lucene 8
Some time around Lucene 7, Lucene changed the migration code to
leave a trace of when the index was originally created. This essentially
makes IndexUpgrader pointless to use, as it no longer upgrades the index
to the newest format.

Instead, we now create a new index and then use addIndexes(CodecReader...)
to add the sub-indexes instead, and use a workaround to make the writer
think that the indexes were created in the newer format. Since this flag
is used for nothing other than the check in IndexWriter, this appears to
be an OK workaround, at least for now.

The strategy for migration is now different for Lucene 7 as well,
as it was found to be misbehaving already. So some indexes which were
already "upgraded" to Lucene 7 will not have to be migrated again.
The logic in VersionGuesser was upgraded for consistency, so such indexes
will now be detected as Lucene 6.